feat: BasicListCell에서 BasicListCellState 분리 및 Converible 프로토콜 정의
